+++ components/actions/notificationActions.ts
@@ -3,7 +3,6 @@
 import { notificationsTable } from "@/db/schema/notifications";
 import { db } from "@/utils/db";
 import { eq, inArray } from "drizzle-orm";
-import { revalidatePath } from "next/cache";
 
 export async function deleteNotification(notificationId: string) {
   try {
@@ -11,7 +10,6 @@ export async function deleteNotification(notificationId: string) {
       .delete(notificationsTable)
       .where(eq(notificationsTable.id, notificationId));
 
-    revalidatePath("/", "layout");
     return { success: true };
   } catch (error) {
     console.error("Delete error: ", error);